requirements-freeze.txt
addict==2.2.0
alabaster==0.7.12
alembic==1.0.5
ansible==2.7.4
apipkg==1.5
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
Babel==2.6.0
banal==0.4.0
bcrypt==3.1.4
bleach==3.0.2
bumpversion==0.5.3
-e git+git@git.unl.edu:helikarlab/ccman.git@50b6e2a84ac59dbfc9ca62f20a9abb3060101179#egg=ccman
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cli-helpers==1.1.0
Click==7.0
click-completion==0.5.0
click-didyoumean==0.0.3
CommonMark==0.5.4
configobj==5.0.6
coverage==4.5.2
cryptography==2.4.2
dataset==1.1.0
decorator==4.3.0
defusedxml==0.5.0
detox==0.18
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
dparse==0.4.1
dropbox==9.2.0
entrypoints==0.2.3
eventlet==0.24.1
execnet==1.5.0
filelock==3.0.10
gitdb2==2.0.5
GitPython==2.1.11
greenlet==0.4.15
-e git+https://github.com/achillesrasquinha/honcho.git@aab83cb10b1d4832c82a4dd3661a6b6df1e1e021#egg=honcho
humanize==0.5.1
idna==2.7
imagesize==1.1.0
ipykernel==5.1.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyrings.alt==3.1
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
ngxtop==0.0.2
normality==0.6.1
notebook==5.7.2
packaging==18.0
pandocfilters==1.4.2
paramiko==2.4.2
pathlib2==2.3.3
pexpect==4.6.0
pgcli==1.10.3
pgspecial==1.11.4
pickleshare==0.7.5
pluggy==0.8.0
prometheus-client==0.5.0
prompt-toolkit==1.0.15
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.3.0
pylint==2.2.2
PyNaCl==1.3.0
pyparsing==2.3.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-forked==0.2
pytest-sugar==0.9.2
pytest-xdist==1.24.1
python-dateutil==2.7.5
python-editor==1.0.3
python-gitlab==1.7.0
pytz==2018.7
PyYAML==4.2b4
pyzmq==17.1.2
qtconsole==4.4.3
recommonmark==0.4.0
redis==3.0.1
requests==2.20.1
rq==0.12.0
safety==1.8.4
semver==2.8.1
Send2Trash==1.5.0
setproctitle==1.1.10
shellingham==1.2.7
simplegeneric==0.8.1
six==1.11.0
slackclient==1.3.0
smmap2==2.0.5
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.14
sqlparse==0.2.4
tabulate==0.8.2
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.4.2
toml==0.10.0
tornado==5.1.1
tox==3.5.3
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==16.1.0
vulture==1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
widgetsnbextension==3.4.2
wrapt==1.10.11
